update imports for schemas split and add coding_request patterns
