+++ tests/graph/test_router_researcher_integration.py
@@ -10,7 +10,7 @@ import pytest
 from check_it_ai.graph.nodes.researcher import researcher_node
 from check_it_ai.graph.nodes.router import router_node
 from check_it_ai.graph.state import AgentState
-from check_it_ai.types.schemas import RouterDecision, RouterTrigger, SearchResult
+from check_it_ai.types import RouterDecision, RouterTrigger, SearchResult
 
 
 class TestRouterResearcherIntegration:
